Display fiat coin prices Calculate total balance from all coins

+++ app/src/main/java/com/blocksdecoded/dex/presentation/widgets/balance/TotalBalanceInfo.kt
@@ -6,5 +6,5 @@ import java.math.BigDecimal
 data class TotalBalanceInfo(
     val coin: Coin,
     val balance: BigDecimal,
-    val fiatBalance: Double
+    val fiatBalance: BigDecimal
 )
